backend/requirements.txt
# Clearpath Chatbot - Backend
fastapi==0.115.5
uvicorn[standard]==0.32.1
groq>=1.0.0
sentence-transformers==3.3.1
faiss-cpu>=1.12.0
pypdf==5.1.0
python-dotenv==1.0.1
pydantic==2.10.2
pydantic-settings==2.6.1
